refactor: add auth_options decorator for CLI commands
